requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4

# S3 Storage
boto3==1.34.34

# Face Recognition and AI
insightface==0.7.3
onnxruntime==1.19.2
opencv-python-headless==4.9.0.80

# Task Queue
celery==5.3.6
redis==5.0.1

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.1.0

# Additional utilities
pillow==10.2.0
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.92.2
